// _backup/backend/src/applications/dto/statut-commission/statut-commission-response.dto.ts
export class StatutCommissionResponseDto {
  id: string;
  code: string;
  nom: string;
  description?: string | null;
  ordreAffichage: number;
  createdAt: Date;
  updatedAt: Date;

  constructor(partial: Partial<StatutCommissionResponseDto>) {
    Object.assign(this, partial);
  }
}